+++ framework/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/common
 
+## 1.2.7
+
+### Patch Changes
+
+- export consolidated retention logic
+
 ## 1.2.6
 
 ### Patch Changes